Update main-3.c
ConvertDec_to_Hex-Oct-Bin.c is a simple conversion program that prompts the user for their name, followed by a prompt for an integer in the range [1 - 1,000,000] inclusive (or “x” to exit the program). The integer entered will be displayed in decimal, hexadecimal, octal, and binary representation. Hexadecimal, octal, and binary representations must be computed algorithmically (i.e. NOT accomplished via format specifier).