Update main-2.c
An integer is divisible by 9 if the sum of its digits is divisible by 9.
Exp: 154,368 1+5+4+3+6+8 = 27 27 is divisible by 9, therefor 154,368 is also divisible by 9. Your program will compute the sum of the individual digits contained within a user-specified integer and will divide this sum by 9 to determine whether or not the (user-specified) integer is divisible by 9.