Update Menue.c
Create a structure type called node with components for the four
integers of an IPv4 address along with a fifth component in which to
store an associated alias (key field) of up to 10 characters. You will then
create a singly linked list of nodes which will contain.
all of the address/alias pairs read in from “CS222_Inet.txt”. For this
exercise, no sorting is required. Once the linked list has been created,
the user will receive the following menu options: